+++ src/main/java/com/irfan/resumebe/Repository/ExperienceRepository.java
@@ -3,5 +3,5 @@ package com.irfan.resumebe.Repository;
 import com.irfan.resumebe.Model.Experience;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExperienceRepository extends JpaRepository<Experience, Long> {
+public interface ExperienceRepository extends JpaRepository<Experience, Integer> {
 }
